Move vm to sandbox folder
