+++ sandbox/vm/legacy_compiler.ts
