+++ sandbox/vm/legacy_gen_index.ts
